list activity on phase change

+++ werewolves.js
@@ -1115,7 +1115,7 @@ client.on("messageReactionAdd", async (reaction, user) => {
 	// Handle reaction ingame
 	} else if(stats.gamephase == gp.INGAME) {
 		// Remove unallowed reactions
-		if(isSpectator(reaction.message.guild.members.cache.get(user.id)) || isDeadParticipant(reaction.message.guild.members.cache.get(user.id)) || isMentor(reaction.message.guild.members.cache.get(user.id))) {
+		if(isSpectator(reaction.message.guild.members.cache.get(user.id)) || isDeadParticipant(reaction.message.guild.members.cache.get(user.id)) || isMentor(reaction.message.guild.members.cache.get(user.id)) || isSub(reaction.message.guild.members.cache.get(user.id)) || isGhost(reaction.message.guild.members.cache.get(user.id))) {
 			if(reaction.emoji == client.emojis.cache.get(stats.no_emoji) || reaction.emoji == client.emojis.cache.get(stats.yes_emoji) || reaction.emoji.name == "🇦" || reaction.emoji.name == "🇧" || reaction.emoji.name == "🇨" || reaction.emoji.name == "🇩" || reaction.emoji.name == "🇪" || reaction.emoji.name == "🇫") return;
 			reaction.users.remove(user);
 		// Automatic pinning
@@ -1138,7 +1138,7 @@ client.on("messageReactionAdd", async (reaction, user) => {
             sql("DELETE FROM prompts WHERE message_id=" + connection.escape(reaction.message.id));
             sql("DELETE FROM action_queue WHERE message_id=" + connection.escape(reaction.message.id));
 			reaction.users.remove(user);
-		}  else if(isGameMaster(reaction.message.guild.members.cache.get(user.id)) && !isParticipant(reaction.message.guild.members.cache.get(user.id)) && reaction.emoji == client.emojis.cache.get(stats.no_emoji)) {
+		}  else if(stats.gamephase == gp.INGAME && isGameMaster(reaction.message.guild.members.cache.get(user.id)) && !isParticipant(reaction.message.guild.members.cache.get(user.id)) && reaction.emoji == client.emojis.cache.get(stats.no_emoji)) {
 			reaction.message.delete();
 		} else if(stats.gamephase == gp.INGAME && isParticipant(reaction.message.guild.members.cache.get(user.id))) {
             let poll = await getPoll(reaction.message.id);
